Refactor navigation and tracing systems
[General]: Introduced `cm_trace_reantrant_state_t` for re-entrant trace functions, enhancing thread safety and performance. Updated related functions to use this structure and return results directly.

[Server]: Added `sv_world_area_mutex` to ensure thread safety in world sector operations. Updated `SV_ClipMoveToEntities`, `SV_Trace`, and `SV_Clip` functions for improved trace handling.

[SVGame]: Refined navigation pathfinding logic in `svg_nav_traversal.cpp` for better edge case handling and compatibility with traversal policies. Enhanced diagnostics and logging for debugging.

[SVGame]: Adjusted sound entity management in `svg_player_weapon.cpp` to optimize allocations and behavior.

[General]: Added new material definitions in `q2rtxp_dev_voxelpaths_dbg_cpuburst.mat`

+++ inc/shared/cm/cm_model.h
@@ -86,15 +86,15 @@ typedef struct cm_s {
 
     // Valid floods and check count:
     int32_t floodValid;
-    int32_t checkCount;
+    int64_t checkCount;
 
     // Null Leaf, as well as null texture, returned in case the query had invalid results
     mleaf_t nullLeaf;
     //mtexinfo_t nullTextureInfo;
 
-    //! Collision Model's bounding box hull.
+    //! Collision Model's shared bounding-box hull template used for initialization and compatibility checks.
     hull_boundingbox_t *hull_boundingbox;
-    //! Collision Model's octagon box hull.
+    //! Collision Model's shared octagon hull template used for initialization and compatibility checks.
     hull_octagonbox_t *hull_octagonbox;
 
     //! Material types, array index equals their typeID. The zero index(0) is used as the default material type.
